Whitesburg Demand Factor Calibration File
This file will display the calibrated demand factors for the Whitesburg system. Instead of manually reproducing the demands within the .inp file itself, once the program has finished running a "temp" file is automatically created that you can save (which will be in the same location as the original file) and will have the factors already in the model.